Blog/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-appconf==1.0.4
django-compressor==2.4
django-npm==1.0.0
djangorestframework==3.11.0
psycopg2==2.8.5
pytz==2020.1
rcssmin==1.0.6
rjsmin==1.1.0
six==1.15.0
sqlparse==0.3.1
